requirements.txt
python-dotenv==1.0.1
pytz==2024.2
supabase==2.9.1
